+++ app/Models/Project.php
@@ -13,7 +13,9 @@ use Spatie\Activitylog\Traits\LogsActivity;
 class Project extends Model
 {
     use HasFactory;
-    use LogsActivity;
+    use LogsActivity {
+        shouldLogEvent as protected shouldLogActivityEvent;
+    }
     use SoftDeletes;
 
     protected static $recordEvents = ['created', 'updated', 'deleted'];
@@ -49,6 +51,17 @@ class Project extends Model
             ])
             ->logOnlyDirty()
             ->dontSubmitEmptyLogs();
+    }
+
+    protected function shouldLogEvent(string $eventName): bool
+    {
+        $causer = auth()->user();
+
+        if (! $causer instanceof User) {
+            return false;
+        }
+
+        return $this->shouldLogActivityEvent($eventName);
     }
 
     public function lessons(): HasMany